+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # master
 
+# 4.1.3 / 2021-04-11
+
+* Add camelCase support for => assignment operator
+* Fix bugs related to is_a? and instance_of?
+
 # 4.1.2 / 2021-04-11
 
 * support => as a right side assignment operator